pyproject.toml
[tool.poetry]
name = "binance-qsmi"
version = "0.0.1"
description = "Binance QSMI"
authors = ["AMiWR <amiwrpremium@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-binance = "^1.0.19"
pydantic = "^2.5.2"
prettytable = "^3.9.0"
typer = "^0.9.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.2"
pylint-pydantic = "^0.3.1"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pytype = "^2023.12.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytype]
inputs = ['binance_qsmi']